requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
coverage==7.8.2
frozenlist==1.7.0
idna==3.10
iniconfig==2.1.0
multidict==6.4.4
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.4.0
python-dotenv==1.1.0
requests==2.32.4
twilio==9.6.3
urllib3==2.4.0
yarl==1.20.1
